# car/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: Car API
    version: 0.0.1
paths:
    /car:
        get:
            tags:
                - Car
            operationId: Car_GetCar
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: name
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GetCarReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /car/create:
        post:
            tags:
                - Car
            operationId: Car_CreateCar
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCarRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreateCarReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /car/helloworld/{name}:
        get:
            tags:
                - Car
            operationId: Car_SayHello
            parameters:
                - name: name
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HelloReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /car/list:
        post:
            tags:
                - Car
            operationId: Car_ListCar
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ListCarRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ListCarReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
    /car/update:
        post:
            tags:
                - Car
            operationId: Car_UpdateCar
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/UpdateCarRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/UpdateCarReply'
                default:
                    description: Default error response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Status'
components:
    schemas:
        CarReply:
            type: object
            properties:
                carId:
                    type: integer
                    format: int64
                carName:
                    type: string
                createTime:
                    type: string
                    format: date-time
                carModel:
                    type: array
                    items:
                        type: string
                carParam:
                    type: object
                    additionalProperties:
                        type: string
        CreateCarReply:
            type: object
            properties:
                car_id:
                    type: integer
                    format: int64
        CreateCarRequest:
            type: object
            properties:
                carName:
                    type: string
                carModel:
                    type: string
                carPhoto:
                    type: string
                    format: bytes
        GetCarReply:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
        GoogleProtobufAny:
            type: object
            properties:
                '@type':
                    type: string
                    description: The type of the serialized message.
            additionalProperties: true
            description: Contains an arbitrary serialized message along with a @type that describes the type of the serialized message.
        HelloReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        ListCarReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/CarReply'
        ListCarRequest:
            type: object
            properties:
                carId:
                    type: integer
                    format: int64
                carName:
                    type: string
                createTime:
                    type: string
                    format: date-time
        Status:
            type: object
            properties:
                code:
                    type: integer
                    description: The status code, which should be an enum value of [google.rpc.Code][google.rpc.Code].
                    format: int32
                message:
                    type: string
                    description: A developer-facing error message, which should be in English. Any user-facing error message should be localized and sent in the [google.rpc.Status.details][google.rpc.Status.details] field, or localized by the client.
                details:
                    type: array
                    items:
                        $ref: '#/components/schemas/GoogleProtobufAny'
                    description: A list of messages that carry the error details.  There is a common set of message types for APIs to use.
            description: 'The `Status` type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by [gRPC](https://github.com/grpc). Each `Status` message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the [API Design Guide](https://cloud.google.com/apis/design/errors).'
        UpdateCarReply:
            type: object
            properties:
                name:
                    type: string
        UpdateCarRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                name:
                    type: string
tags:
    - name: Car